requirements.txt
Faker==19.3.1
selenium==4.13.0
requests==2.31.0
webdriver-manager==4.0.0
pytest==8.0.2
pytest-xdist==3.5.0
pytest-html==4.1.1
allure-pytest==2.13.2